updated cheat mode
